Move inviteToTestBSON into an online test specific trait
